refactor(ktf): inline define_class_java, drop classfile module
